重構meta tag的程式碼 並加入新創計劃的meta tag
新創計劃的meta tag會在title加入 (新創計劃) 字樣以跟已成立的公司區分
description中會顯示投資截止時間

format時間的function是直接從client複製過來的
