+++ server/startup/metaTag/createMeta.js
@@ -1,0 +1,7 @@
+export function createMetaProperty(property, content) {
+  return `<meta property="${property}" content="${content}" />`;
+}
+
+export function createMetaName(name, content) {
+  return `<meta name="${name}" content="${content}" />`;
+}
